requirements.txt
fastapi>=0.135.2
uvicorn[standard]>=0.42.0
httpx>=0.28.1
python-dotenv>=1.2.2
PyYAML>=6.0.3
pandas>=3.0.1
boto3>=1.42.68